[common/functions] add default impl for IFunction
