+++ common/functions/src/comparisons/comparison.rs
@@ -4,11 +4,10 @@
 
 use std::fmt;
 
-use anyhow::{bail, ensure, Result};
+use anyhow::{ensure, Result};
 use common_datablocks::DataBlock;
 use common_datavalues::{
-    self as datavalues, DataColumnarValue, DataSchema, DataType, DataValue,
-    DataValueComparisonOperator,
+    self as datavalues, DataColumnarValue, DataSchema, DataType, DataValueComparisonOperator,
 };
 
 use crate::comparisons::{
@@ -81,29 +80,6 @@ impl IFunction for ComparisonFunction {
 
     fn set_depth(&mut self, depth: usize) {
         self.depth = depth;
-    }
-
-    fn accumulate(&mut self, block: &DataBlock) -> Result<()> {
-        self.left.accumulate(block)?;
-        self.right.accumulate(block)
-    }
-
-    fn accumulate_result(&self) -> Result<Vec<DataValue>> {
-        bail!(
-            "Unsupported accumulate_result operation for function {}",
-            self.op
-        );
-    }
-
-    fn merge(&mut self, _states: &[DataValue]) -> Result<()> {
-        bail!("Unsupported merge operation for function {}", self.op);
-    }
-
-    fn merge_result(&self) -> Result<DataValue> {
-        bail!(
-            "Unsupported merge_result operation for function {}",
-            self.op
-        );
     }
 }
