+++ common/functions/src/function.rs
@@ -4,7 +4,7 @@
 
 use std::fmt;
 
-use anyhow::Result;
+use anyhow::{bail, Result};
 use common_datablocks::DataBlock;
 use common_datavalues::{DataColumnarValue, DataSchema, DataType, DataValue};
 use dyn_clone::DynClone;
@@ -13,11 +13,25 @@ pub trait IFunction: fmt::Display + Sync + Send + DynClone {
     fn return_type(&self, input_schema: &DataSchema) -> Result<DataType>;
     fn nullable(&self, input_schema: &DataSchema) -> Result<bool>;
     fn eval(&self, block: &DataBlock) -> Result<DataColumnarValue>;
-    fn set_depth(&mut self, depth: usize);
-    fn accumulate(&mut self, block: &DataBlock) -> Result<()>;
-    fn accumulate_result(&self) -> Result<Vec<DataValue>>;
-    fn merge(&mut self, states: &[DataValue]) -> Result<()>;
-    fn merge_result(&self) -> Result<DataValue>;
+
+    fn set_depth(&mut self, _depth: usize) {}
+
+    fn accumulate(&mut self, _block: &DataBlock) -> Result<()> {
+        bail!("Function Error: accumulate unimplemented");
+    }
+
+    fn accumulate_result(&self) -> Result<Vec<DataValue>> {
+        bail!("Function Error: accumulate_result unimplemented");
+    }
+
+    fn merge(&mut self, _states: &[DataValue]) -> Result<()> {
+        bail!("Function Error: merge unimplemented");
+    }
+
+    fn merge_result(&self) -> Result<DataValue> {
+        bail!("Function Error: merge_result unimplemented");
+    }
+
     fn is_aggregator(&self) -> bool {
         false
     }
